Maintain backward compatible
