+++ omisego-sdk/src/main/java/co/omisego/omisego/websocket/listener/SocketCustomEventListener.kt
@@ -7,12 +7,16 @@ package co.omisego.omisego.websocket.listener
  * Copyright © 2017-2018 OmiseGO. All rights reserved.
  */
 
+import co.omisego.omisego.model.APIError
 import co.omisego.omisego.model.socket.SocketReceive
+import co.omisego.omisego.model.transaction.consumption.TransactionConsumption
 import co.omisego.omisego.websocket.event.SocketEvent
+import co.omisego.omisego.websocket.event.TransactionConsumptionFinalizedEvent
+import co.omisego.omisego.websocket.event.TransactionConsumptionRequestEvent
 import co.omisego.omisego.websocket.strategy.FilterStrategy
 
 interface SocketCustomEventListener {
-    abstract var strategy: FilterStrategy
+    var strategy: FilterStrategy
     fun onEvent(event: SocketEvent<*>)
 
     companion object {
@@ -55,6 +59,50 @@ interface SocketCustomEventListener {
                     lambda(event)
                 }
             }
+        }
+    }
+    abstract class TransactionRequestListener : SimpleSocketCustomEventListener<SocketEvent<*>>() {
+        override var strategy: FilterStrategy = FilterStrategy.Event(allowedEvents)
+        final override fun onSpecificEvent(event: SocketEvent<*>) {
+            when (event) {
+                is TransactionConsumptionRequestEvent -> event.socketReceive.data?.let(::onTransactionConsumptionRequest)
+                is TransactionConsumptionFinalizedEvent -> event.socketReceive.dispatch(
+                    onSuccess = ::onTransactionConsumptionFinalizedSuccess,
+                    onError = ::onTransactionConsumptionFinalizedFail
+                )
+            }
+        }
+
+        abstract fun onTransactionConsumptionRequest(transactionConsumption: TransactionConsumption)
+        abstract fun onTransactionConsumptionFinalizedSuccess(transactionConsumption: TransactionConsumption)
+        abstract fun onTransactionConsumptionFinalizedFail(transactionConsumption: TransactionConsumption, apiError: APIError)
+
+        companion object {
+            internal val allowedEvents = listOf(
+                TransactionConsumptionRequestEvent::class.java,
+                TransactionConsumptionFinalizedEvent::class.java
+            )
+        }
+    }
+
+    abstract class TransactionConsumptionListener : SimpleSocketCustomEventListener<SocketEvent<*>>() {
+        override var strategy: FilterStrategy = FilterStrategy.Event(allowedEvents)
+        final override fun onSpecificEvent(event: SocketEvent<*>) {
+            when (event) {
+                is TransactionConsumptionFinalizedEvent -> event.socketReceive.dispatch(
+                    onSuccess = ::onTransactionConsumptionFinalizedSuccess,
+                    onError = ::onTransactionConsumptionFinalizedFail
+                )
+            }
+        }
+
+        abstract fun onTransactionConsumptionFinalizedSuccess(transactionConsumption: TransactionConsumption)
+        abstract fun onTransactionConsumptionFinalizedFail(transactionConsumption: TransactionConsumption, apiError: APIError)
+
+        companion object {
+            private val allowedEvents = listOf(
+                TransactionConsumptionFinalizedEvent::class.java
+            )
         }
     }
 }
